Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Warning

This form of benchmarking doesn't currently work for Threaded()and you'll likely get lot's of errors. See Thread Timers for an alternative.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:            18.7s / 0.19%           1.64GiB / 0.39%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2   35.0ms   100%  17.5ms   6.21MiB  94.6%  3.10MiB
   applying f to su...  2.00k   21.8ms  61.9%  10.9μs   4.02MiB  61.3%  2.06KiB
   submatrix constr...  2.00k   10.9ms  31.0%  5.44μs   2.19MiB  33.3%  1.12KiB
   writing result c...  2.00k    301μs  0.86%   151ns     0.00B  0.00%    0.00B
 allocate result            2    109μs  0.31%  54.4μs    360KiB  5.35%   180KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Note

You can safely call all *_benchmarks functions in combination with @everywhere, i.e. when using multiple Julia workers (Distributed.jl).

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.07733076
2: 0.641658971
3: 0.68060306
4: 0.659436276
5: 0.65314164
6: 0.656831043
7: 0.661585097
8: 0.655825814
9: 0.659082952
10: 0.648167761

Total (sum): 5.993663374
Std (σ): 0.18370208026964888

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.117538884    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.11521505     
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.117549916    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.117475909    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.120081245   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.116744919    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.116668       
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.117308384    
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.116005124    
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.113479532    
      └                                        ┘ 
Total (sum): 1.1680669629999998
Std (σ): 0.001724967349108001
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.172182068   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.171932199   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.172806443   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.170482141   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165198174    
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.8526010250000001
Std (σ): 0.0030946672968392808

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1102    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1102    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1109    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1101    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1119    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1073     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1022      
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1136   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1067     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1057     
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▏ 3                                       
   ( 70.0,  75.0] ┤▌ 29                                      
   ( 75.0,  80.0] ┤██▎ 123                                   
   ( 80.0,  85.0] ┤██████▌ 387                               
   ( 85.0,  90.0] ┤█████████████▌ 798                        
   ( 90.0,  95.0] ┤██████████████████████████▍ 1558          
   ( 95.0, 100.0] ┤████████████████████████████████▋ 1932    
   (100.0, 105.0] ┤██████████████████████████████████  2007  
   (105.0, 110.0] ┤████████████████████████▋ 1456            
   (110.0, 115.0] ┤████████████████▏ 945                     
   (115.0, 120.0] ┤███████▊ 464                              
   (120.0, 125.0] ┤███▋ 215                                  
   (125.0, 130.0] ┤▉ 58                                      
   (130.0, 135.0] ┤▍ 23                                      
   (135.0, 140.0] ┤▏ 1                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.print_timings

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_timings — Method

Pretty print the given timings.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks