Launch Configurations

Index

SubmatrixMethod.LaunchConfiguration

SubmatrixMethod.Serial

SubmatrixMethod.SerialMPI

SubmatrixMethod.Threaded

References

SubmatrixMethod.LaunchConfiguration — Type

An abstract launch configuration for submatrix_apply. See subtypes(LaunchConfiguration) for a list of available launch configurations.

SubmatrixMethod.Serial — Type

No parallelism. (BLAS might be parallelized, of course.)

Fields / keyword arguments:

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

SubmatrixMethod.SerialMPI — Type

Use multiprocessing via MPI. On each MPI worker, run things in serial.

Fields / keyword arguments:

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

SubmatrixMethod.Threaded — Type

Multithreading.

Fields / keyword arguments:

threads (default: false): toggles composable, load-balanced multithreading (based on Threads.@spawn) such that different submatrices are created and processed on different threads.

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

taskbased (default: true): if true, we use @spawn, i.e. tasks and dynamic load balancing. If false, we use @threads :static.