Submatrix Method: API

Index

SubmatrixMethod.submatrix_apply

SubmatrixMethod.submatrix_apply

SubmatrixMethod.HDF5Input

References

SubmatrixMethod.HDF5Input — Type

Represents a matrix in an HDF5 File that has been stored with SubmatrixMethod.write_hdf5.

Fields / keyword arguments:

fname: filename / path to file

path: path to group that contains the matrix inside the file

mmap (default: false): toggle memory mapping

SubmatrixMethod.submatrix_apply — Method

submatrix_apply(f, A[, lc=Serial()])

Uses the submatrix method to approximately compute f(A), i.e. the application of the matrix function f to the input matrix A.

The input matrix A must be

positive definite (check with isposdef),

symmetric (check with issymmetric).

and should be

sparse (check with sparsity or density).

A LaunchConfiguration, such as Serial or Threaded, which specifies the modus operandi can be provided as a third argument (lc).

SubmatrixMethod.submatrix_apply — Method

submatrix_apply(f, file::HDF5Input, [, lc=Serial()])

Reads the input matrix A from the given file::HDF5Input and calls submatrix_apply(f, A, lc) afterwards.

See HDF5Input.