IO

Writing, reading, and memory mapping SparseMatrixCSC to and from disk.

HDF5

Functionality for writing, reading, or memory mapping a SparseMatrixCSC[Tv, Ti} into or from a specific path in a HDF5 file. For memory mapping, the value type (Tv) must be isbitstype (e.g Float64, Float32. or ComplexF64).

We'll just write / read the fields of the SparseMatrixCSC into / from separate datasets under the given path.

Example:

julia> S = sprand(100, 100, 0.001);
julia> SubmatrixMethod.write_hdf5("mymatrix.h5", "S", S)
julia> isfile("mymatrix.h5")true
julia> T = SubmatrixMethod.read_hdf5("mymatrix.h5", "S");
julia> S == Ttrue
julia> R = SubmatrixMethod.read_hdf5("mymatrix.h5", "S"; mmap=true);
julia> R == Ttrue

Single binary file

We define our own Binary Compressed Sparse Column (BCSC) format for storing a SparseMatrixCSC{Tv, Ti} on disk. The value type Tv must be a regular floating point type (e.g. Float64 or Float32) and the index type Ti must be a regular integer type (e.g. Int64 or Int128).

"Definition" of the format based on the SparseMatrixCSC data structure:

# header
write(f, sizeof(Tv))
write(f, sizeof(Ti))
write(f, length(S.colptr))
write(f, length(S.rowval))
write(f, length(S.nzval))
# data
write(f, S.m)
write(f, S.n)
write(f, S.colptr)
write(f, S.rowval)
write(f, S.nzval)

Example:

julia> S = sprand(100, 100, 0.001);
julia> SubmatrixMethod.write_bcsc("mymatrix.bcsc", S)
julia> isfile("mymatrix.bcsc")true
julia> T = SubmatrixMethod.read_bcsc("mymatrix.bcsc");
julia> S == Ttrue
julia> R = SubmatrixMethod.read_bcsc("mymatrix.bcsc"; mmap=true);
julia> R == Ttrue

Warning

As for now, we don't care about endianness etc. which means that, in general, our format is not portable! As an alternative, consider using HDF5 instead.

Three binary files

The fields colptr, rowval, and nzval of the SparseMatrixCSC are simply stored in separate binary files with the file extensions .colptr, .rowval, and .nzval, respectively.

Example:

julia> S = sprand(100, 100, 0.001);
julia> SubmatrixMethod.write_three_files("mymatrix", S)
julia> filter(startswith("mymatrix"), readdir())3-element Vector{String}:
 "mymatrix.colptr"
 "mymatrix.nzval"
 "mymatrix.rowval"
julia> T = SubmatrixMethod.read_three_files("mymatrix");
julia> S == Ttrue
julia> R = SubmatrixMethod.read_three_files("mymatrix"; mmap=true);
julia> R == Ttrue

Index

SubmatrixMethod.read_bcsc

SubmatrixMethod.read_hdf5

SubmatrixMethod.read_three_files

SubmatrixMethod.read_txt

SubmatrixMethod.write_bcsc

SubmatrixMethod.write_hdf5

SubmatrixMethod.write_three_files

References

SubmatrixMethod.read_bcsc — Method

Creates a S::SparseMatrixCSC from a binary file in BCSC format, e.g. created using write_bcsc.

Keyword arguments:

mmap (default: false): If true, the data of the arrays colptr, rowval, and nzval within the returned SparseMatrixCSC is linked to the data in the binary file on disk (memory mapping).

See: write_bcsc

SubmatrixMethod.read_hdf5 — Method

Creates a S::SparseMatrixCSC from the dataset at path in the HDF5 file fname, which should have been created using write_hdf5.

Keyword arguments:

mmap (default: false): If true, the data of the arrays colptr, rowval, and nzval within the returned SparseMatrixCSC is linked to the data in the HDF5 file on disk (memory mapping).

See: write_hdf5

SubmatrixMethod.read_three_files — Function

Creates a S::SparseMatrixCSC by reading the binary data of the fields from the files prefix.colptr, prefix.rowval, and prefix.nzval.

Important note: This functions assumes that the sparse matrix is square, i.e. that m == n!

Keyword arguments:

mmap (default: false): If true, the data of the arrays colptr, rowval, and nzval within the returned SparseMatrixCSC is linked to the respective binary files on disk (memory mapping).

See: write_three_files

SubmatrixMethod.read_txt — Function

Read in txt files produced by Matlab (see Michael's original implementation) and return a SparseMatrixCSC.

SubmatrixMethod.write_bcsc — Method

Writes a SparseMatrixCSC to file in our custom BCSC (Binary Compressed Sparse Column) format.

See: read_bcsc

SubmatrixMethod.write_hdf5 — Method

Writes a SparseMatrixCSC into the HDF5 file fname under path. Will create the file if it doesn't exist yet.

See: read_hdf5

SubmatrixMethod.write_three_files — Method

Writes S::SparseMatrixCSC to three binary files prefix.colptr, prefix.rowval, and prefix.nzval storing the respective fieldvalues of the sparse matrix.

See: read_three_files