Launch Configurations

Index

SubmatrixMethod.DistributedSerial

SubmatrixMethod.LaunchConfiguration

SubmatrixMethod.MPISerial

SubmatrixMethod.Serial

SubmatrixMethod.Threaded

References

SubmatrixMethod.DistributedSerial — Type

Use multiprocessing via Distributed.jl. On each Julia worker, run things in serial.

Fields / keyword arguments:

nworkers (default: Distributed.nworkers()): the workers workers()[1:nworkers] will be used in the computation. Distributed.nworkers() is the default and an upper bound.

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

balance (default: true): if true, we distribute submatrices as evenly as possible (w.r.t. their size) across Julia workers.

show_distribution (default: false): if true, we print the distribution of submatrices and their weight (i.e. size) across Julia workers. If UnicodePlots.jl is loaded, bar plots are produced.

SubmatrixMethod.LaunchConfiguration — Type

An abstract launch configuration for submatrix_apply. See subtypes(LaunchConfiguration) for a list of available launch configurations.

SubmatrixMethod.MPISerial — Type

Use multiprocessing via MPI. On each MPI worker, run things in serial.

When using this configuration, the result of submatrix_apply is only returned on the master process whereas nothing is returned on all other MPI processes (workers).

Fields / keyword arguments:

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

verbose (default: false): if true, outputs MPI communication meta information etc.

balance (default: true): if true, we distribute submatrices as evenly as possible (w.r.t. their size) across MPI workers.

show_distribution (default: false): if true, we print the distribution of submatrices and their weight (i.e. size) across MPI workers. If UnicodePlots.jl is loaded, bar plots are produced.

SubmatrixMethod.Serial — Type

No parallelism. (BLAS might be parallelized, of course.)

Fields / keyword arguments:

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

SubmatrixMethod.Threaded — Type

Multithreading.

Fields / keyword arguments:

mode (default: :threads): If :spawn, @spawn is used, i.e. tasks and dynamic load balancing. If :batch, Polyester's @batch is used, i.e. no load balancing. Else, @threads :static is used, i.e. no load balancing.

nthreads (default: Threads.nthreads()): For some modes, this allows one to set the number of used Julia threads. Threads.nthreads() is the default and an upper bound.

balance (default: false): if true, we distribute submatrices as evenly as possible (w.r.t. their size) across threads. Not available for mode = :spawn.

show_distribution (default: false): if true, we print the distribution of submatrices and their weight (i.e. size) across threads. If UnicodePlots.jl is loaded, bar plots are produced.

inplace (default: false): if true, expects that f is an in-place matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.