Submatrix Method

Index

SubmatrixMethod.construct_submatrix

SubmatrixMethod.submatrix_apply

SubmatrixMethod.submatrix_computation!

SubmatrixMethod.submatrix_computation_inplace!

References

SubmatrixMethod.construct_submatrix — Method

construct_submatrix(A, j) -> submatrix, indices

Construct the j-th submatrix form the input matrix A.

SubmatrixMethod.submatrix_apply — Function

submatrix_apply(f, A[, lc])

Uses the submatrix method to approximately compute f(A), i.e. the application of the matrix function f to the input matrix A.

The input matrix A must be

positive definite (check with isposdef),

symmetric (check with issymmetric).

and should be

sparse (check with sparsity or density).

A LaunchConfiguration, such as Serial or Threaded, which specifies the modus operandi can be provided as a third argument (lc).

SubmatrixMethod.submatrix_computation! — Method

submatrix_computation!(R, f, A, j)

This is the "computation kernel". Constructs the j-th submatrix from the input matrix A, applies the matrix function f to it, and fills the corresponding column of the result matrix R.

SubmatrixMethod.submatrix_computation_inplace! — Method

submatrix_computation_inplace!(R, f_inplace, A, j)

This is the "computation kernel". Constructs the j-th submatrix from the input matrix A, applies the matrix function f_inplace to it (in-place), and fills the corresponding column of the result matrix R.