Utility

Index

SubmatrixMethod.density

SubmatrixMethod.inv!

SubmatrixMethod.sparsity

SubmatrixMethod.spectral_norm

SubmatrixMethod.sprandsymposdef

SubmatrixMethod.sprandsymposdef

References

SubmatrixMethod.density — Method

Number of non-zero elements over the total number of elements.

SubmatrixMethod.inv! — Method

inv!(M)

In-place matrix inversion. Overwrites the input matrix M with the result.

SubmatrixMethod.sparsity — Method

Number of zero-valued elements over the total number of elements.

SubmatrixMethod.spectral_norm — Method

Computes the spectral norm of the input matrix:

\[\|A\|_{2}=\sqrt{\lambda_{\max }\left(A^{H} A\right)},\]

i.e. the square root of the maximal eigenvalue of A'*A.

SubmatrixMethod.sprandsymposdef — Method

sprandsymposdef(n,p,c)

Generates a sparse matrix of size (n,n) with the following properties:

issymmetric

isposdef

has approximately p * n^2 non-zero elements

has a condition number of exactly c

WARNING: While it works okish for some inputs it can take ages to converge (if at all) for others!

Algorithm:

Creates a diagonal positive-definite matrix with the desired condition number and applies random Jacobi rotations to it to create non-zero off-diagonal elements.

References:

The function is inspired by MATLABs sprandsym(n,density,rc,kind=1).

SubmatrixMethod.sprandsymposdef — Method

sprandsymposdef(n,p)

Generates a sparse matrix of size (n,n) with the following properties:

issymmetric

isposdef

has approximately p * n^2 non-zero elements

has a condition number of approximately c ≈ 1.

Algorithm:

Creates a random sparse matrix (sprandn), symmetrizes it and makes it diagonally dominant by adding n*I to ensure positive definiteness.

References:

The implementation of the function is inspired by this stackexchange thread.