Submatrix Method

Index

SubmatrixMethod.construct_submatrix

SubmatrixMethod.submatrix_apply

SubmatrixMethod.submatrix_computation!

SubmatrixMethod.submatrix_computation_inplace!

References

SubmatrixMethod.construct_submatrix — Method

construct_submatrix(A, j) -> submatrix, indices

Construct the j-th submatrix form the input matrix A.

SubmatrixMethod.submatrix_apply — Method

submatrix_apply(f, A)

Uses the submatrix method to approximately compute f(A), i.e. the application of the matrix function f to the input matrix A.

The input matrix A

must be symmetric and positive definite (check with isposdef),

should be sparse (check with sparsity).

The keyword argument multithreading can be used to enable parallel processing of different submatrices. The available options are:

false: no multithreading (default)

true: composable, load-balanced multithreading (Threads.@spawn)

Further keyword arguments:

inplace (default: false): if true, expects that f is an in-place

matrix function that overwrites the input matrix with the result, i.e. inv! instead of inv, for example.

SubmatrixMethod.submatrix_computation! — Method

submatrix_computation!(X, f, A, j)

This is the "computation kernel". Constructs the j-th submatrix from the input matrix A, applies the matrix function f to it, and fills the corresponding column of the result matrix X.

SubmatrixMethod.submatrix_computation_inplace! — Method

submatrix_computation_inplace!(X, f_inplace, A, j)

This is the "computation kernel". Constructs the j-th submatrix from the input matrix A, applies the matrix function f_inplace to it (in-place), and fills the corresponding column of the result matrix X.