Utility

Index

SubmatrixMethod.density

SubmatrixMethod.generate_input_matrix

SubmatrixMethod.inv!

SubmatrixMethod.sparsity

SubmatrixMethod.spectral_norm

References

SubmatrixMethod.density — Method

Number of non-zero elements over the total number of elements.

SubmatrixMethod.inv! — Method

inv!(M)

In-place matrix inversion. Overwrites the input matrix M with the result.

SubmatrixMethod.sparsity — Method

Number of zero-valued elements over the total number of elements.

SubmatrixMethod.spectral_norm — Method

Computes the spectral norm of the input matrix:

\[\|A\|_{2}=\sqrt{\lambda_{\max }\left(A^{H} A\right)},\]

i.e. the square root of the maximal eigenvalue of A'*A.

SubmatrixMethod.generate_input_matrix — Method

generate_input_matrix(n,density)

Generate a sparse n x n symmetric, positive definite matrix with a density of non-zero elements of approximately density.

TODO: Improve this so that we can control the condition number. TODO: Compare to MATLABs sprandsym(size,density,1/condition,kind) with kind=1

Inspired by

https://www.mathworks.com/help/matlab/ref/sprandsym.html

https://stackoverflow.com/questions/46930325/create-a-sparse-symmetric-random-matrix-in-julia