Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:           47916s / 0.00%           1.55GiB / 3.44%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    580ms   100%   290ms   54.2MiB  99.4%  27.1MiB
   applying f to su...  2.00k    165ms  28.4%  82.5μs   3.94MiB  7.22%  2.02KiB
   submatrix constr...  2.00k   10.5ms  1.81%  5.24μs   2.14MiB  3.92%  1.09KiB
   writing result c...  2.00k    304μs  0.05%   152ns     0.00B  0.00%    0.00B
 allocate result            2   53.0μs  0.01%  26.5μs    354KiB  0.63%   177KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.120463257
2: 0.694293728
3: 0.729914199
4: 0.722317538
5: 0.72508279
6: 0.692478906
7: 0.710187867
8: 0.692358416
9: 0.704748344
10: 0.714412286

Total (sum): 6.5062573310000005
Std (σ): 0.18678219874731178

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■ 0.082258579                           
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.674190308  
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.67718009   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.675374414  
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.664011923   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.643690821   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.667839324  
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.662986221   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.665917347  
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.630352884    
      └                                        ┘ 
Total (sum): 6.043801910999998
Std (σ): 0.18404372514971248
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■ 0.214535078                      
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.691107096   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.632489679     
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.675356104    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.633966485     
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 2.847454442
Std (σ): 0.20007343860177323

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1125   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1075     
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1052     
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1089    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1092    
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1080    
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1095    
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1064     
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1086   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▎ 5                                       
   ( 70.0,  75.0] ┤▋ 34                                      
   ( 75.0,  80.0] ┤██▌ 146                                   
   ( 80.0,  85.0] ┤██████▉ 399                               
   ( 85.0,  90.0] ┤███████████████▍ 888                      
   ( 90.0,  95.0] ┤██████████████████████████▋ 1534          
   ( 95.0, 100.0] ┤██████████████████████████████████  1959  
   (100.0, 105.0] ┤█████████████████████████████████▌ 1927   
   (105.0, 110.0] ┤█████████████████████████▌ 1465           
   (110.0, 115.0] ┤███████████████▊ 912                      
   (115.0, 120.0] ┤███████▋ 445                              
   (120.0, 125.0] ┤███▌ 199                                  
   (125.0, 130.0] ┤█▎ 64                                     
   (130.0, 135.0] ┤▍ 19                                      
   (135.0, 140.0] ┤▏ 2                                       
   (140.0, 145.0] ┤▏ 1                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks