Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Warning

This form of benchmarking doesn't currently work for Threaded()and you'll likely get lot's of errors. See Thread Timers for an alternative.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:            14.8s / 0.39%           1.63GiB / 0.41%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2   57.8ms   100%  28.9ms   6.46MiB  94.8%  3.23MiB
   applying f to su...  2.00k   47.1ms  81.3%  23.5μs   4.27MiB  62.6%  2.18KiB
   submatrix constr...  2.00k   8.33ms  14.4%  4.16μs   2.19MiB  32.2%  1.12KiB
   writing result c...  2.00k    301μs  0.52%   151ns     0.00B  0.00%    0.00B
 allocate result            2   84.2μs  0.15%  42.1μs    359KiB  5.15%   180KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Note

You can safely call all *_benchmarks functions in combination with @everywhere, i.e. when using multiple Julia workers (Distributed.jl).

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.076089233
2: 0.075884442
3: 0.075015585
4: 0.075353612
5: 0.076794156
6: 0.075490954
7: 0.074974006
8: 0.074012348
9: 0.07589602
10: 0.07606118

Total (sum): 0.7555715359999999
Std (σ): 0.0007713529063264544

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.071609758   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.070449437    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.071128874   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.069603581    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.072317691   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.07080857     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.069741861    
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.068721375    
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.07083572     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.07182681    
      └                                        ┘ 
Total (sum): 0.707043677
Std (σ): 0.0011068542885013328
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.146618146   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.14444116    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.146252357   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.142254543    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.146213234   
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.7257794399999999
Std (σ): 0.001829416724891427

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1083    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1130   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1052     
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1054     
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1091    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1079     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1060     
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1125   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1108    
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1086   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1087   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 65.0,  70.0] ┤▏ 4                                       
   ( 70.0,  75.0] ┤▌ 34                                      
   ( 75.0,  80.0] ┤█▉ 116                                    
   ( 80.0,  85.0] ┤██████▊ 400                               
   ( 85.0,  90.0] ┤██████████████▊ 866                       
   ( 90.0,  95.0] ┤█████████████████████████▎ 1470           
   ( 95.0, 100.0] ┤███████████████████████████████▊ 1862     
   (100.0, 105.0] ┤██████████████████████████████████  1987  
   (105.0, 110.0] ┤██████████████████████████▊ 1573          
   (110.0, 115.0] ┤███████████████▋ 919                      
   (115.0, 120.0] ┤████████▍ 491                             
   (120.0, 125.0] ┤███▍ 194                                  
   (125.0, 130.0] ┤▉ 57                                      
   (130.0, 135.0] ┤▍ 19                                      
   (135.0, 140.0] ┤▎ 7                                       
   (140.0, 145.0] ┤▏ 1                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.print_timings

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_timings — Method

Pretty print the given timings.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks