Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:             153s / 0.31%           1.54GiB / 3.47%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    472ms   100%   236ms   54.5MiB  99.4%  27.3MiB
   applying f to su...  2.00k   54.1ms  11.5%  27.0μs   4.17MiB  7.61%  2.14KiB
   submatrix constr...  2.00k   9.35ms  1.98%  4.67μs   2.19MiB  3.99%  1.12KiB
   writing result c...  2.00k    300μs  0.06%   150ns     0.00B  0.00%    0.00B
 allocate result            2   47.5μs  0.01%  23.7μs    358KiB  0.64%   179KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.93704359
2: 1.052677783
3: 1.085704736
4: 1.075916692
5: 1.089004675
6: 1.068392075
7: 1.084587655
8: 1.092020128
9: 1.093485829
10: 1.081338037

Total (sum): 10.6601712
Std (σ): 0.046953952460562644

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.624291875     
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.687586767   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.67788672    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.674879835   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.67526625    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.656066846    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.632025693     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.634471168     
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.63956115      
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■ 0.614210505      
      └                                        ┘ 
Total (sum): 6.516246808999999
Std (σ): 0.02597695579801344
julia> submatrix_apply(inv, A, Threaded(timers=true, mode=:batch, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.871858419   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■ 0.766661397      
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.80417941      
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■ 0.773596071      
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.793404194     
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 4.009699491
Std (σ): 0.0418760623668646

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1076      
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1129    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098     
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1131    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1105     
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1119     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1171   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1057      
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1096     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1082      
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1106   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1105   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1106   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1106   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1106   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 65.0,  70.0] ┤▎ 5                                       
   ( 70.0,  75.0] ┤▌ 29                                      
   ( 75.0,  80.0] ┤██▍ 129                                   
   ( 80.0,  85.0] ┤██████▊ 387                               
   ( 85.0,  90.0] ┤███████████████▎ 865                      
   ( 90.0,  95.0] ┤█████████████████████████▋ 1465           
   ( 95.0, 100.0] ┤█████████████████████████████████▌ 1917   
   (100.0, 105.0] ┤██████████████████████████████████  1942  
   (105.0, 110.0] ┤██████████████████████████▋ 1520          
   (110.0, 115.0] ┤████████████████▉ 968                     
   (115.0, 120.0] ┤████████▊ 500                             
   (120.0, 125.0] ┤███▍ 186                                  
   (125.0, 130.0] ┤█▎ 64                                     
   (130.0, 135.0] ┤▍ 19                                      
   (135.0, 140.0] ┤▏ 4                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks