Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with multithreading=false (default), can be benchmarked in this way. For multithreading=true, you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = SubmatrixMethod.generate_input_matrix(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:             971s / 0.04%           1.50GiB / 4.37%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    343ms  89.6%   172ms   50.5MiB  75.0%  25.2MiB
   applying f to su...  2.00k   30.4ms  7.95%  15.2μs   13.9MiB  20.7%  7.13KiB
   submatrix constr...  2.00k   9.56ms  2.50%  4.78μs   2.32MiB  3.45%  1.19KiB
   writing result c...  2.00k    560μs  0.15%   280ns     0.00B  0.00%    0.00B
 allocate result            2   39.7ms  10.4%  19.9ms   16.8MiB  25.0%  8.39MiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, there is a thread_timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = SubmatrixMethod.generate_input_matrix(1000, 0.1);
julia> submatrix_apply(inv, A; multithreading=true, thread_timers=true);
Individual thread timings

1: 0.29453554899999995
2: 0.3003050440000001
3: 0.29136652300000015
4: 0.296509178
5: 0.298800244
6: 0.294573159
7: 0.28924174199999997
8: 0.283643049
9: 0.29143162399999994
10: 0.2878254169999999

Total (sum): 2.9282315290000005

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A; multithreading=true, thread_timers=true);
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■ 0.3360950629999999   
    2 ┤■■■■■■■■■■■■■■■■■■■■ 0.33793394799999993  
    3 ┤■■■■■■■■■■■■■■■■■■■■ 0.3382220770000001   
    4 ┤■■■■■■■■■■■■■■■■■■■■ 0.33789079699999996  
    5 ┤■■■■■■■■■■■■■■■■■■■■ 0.33743481700000005  
    6 ┤■■■■■■■■■■■■■■■■■■■■ 0.33772407900000007  
    7 ┤■■■■■■■■■■■■■■■■■■■■ 0.33636613599999987  
    8 ┤■■■■■■■■■■■■■■■■■■■■ 0.3362105740000001   
    9 ┤■■■■■■■■■■■■■■■■■■■■ 0.3371072019999999   
   10 ┤■■■■■■■■■■■■■■■■■■■■ 0.33695216000000006  
      └                                        ┘ 
Total (sum): 3.371936853

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks