Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:        18446410402s / 0.00%        1.47GiB / 4.13%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    401ms   100%   200ms   61.8MiB  99.4%  30.9MiB
   applying f to su...  2.00k   13.6ms  3.38%  6.78μs   13.6MiB  21.9%  6.95KiB
   submatrix constr...  2.00k   9.04ms  2.26%  4.52μs   2.22MiB  3.58%  1.14KiB
   writing result c...  2.00k    307μs  0.08%   153ns     0.00B  0.00%    0.00B
 allocate result            2   45.0μs  0.01%  22.5μs    363KiB  0.57%   181KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a thread_timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(thread_timers=true));
Individual thread timings

1: 0.10162365400000002
2: 0.10803387800000001
3: 0.08610122500000002
4: 0.10467458500000008
5: 0.10532374900000004
6: 0.08945043400000004
7: 0.095934793
8: 0.10700549899999998
9: 0.10268054399999997
10: 0.09853263399999995

Total (sum): 0.9993609950000002

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(thread_timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■ 0.08466361700000005   
    2 ┤■■■■■■■■■■■■■■■■■■■ 0.08535553199999996   
    3 ┤■■■■■■■■■■■■■■■■■■■ 0.08541390399999998   
    4 ┤■■■■■■■■■■■■■■■■■■■ 0.085800534           
    5 ┤■■■■■■■■■■■■■■■■■■■ 0.08576952200000001   
    6 ┤■■■■■■■■■■■■■■■■■■■ 0.08610908299999999   
    7 ┤■■■■■■■■■■■■■■■■■■■ 0.08547133700000001   
    8 ┤■■■■■■■■■■■■■■■■■■■ 0.08563062000000002   
    9 ┤■■■■■■■■■■■■■■■■■■■ 0.08533820200000003   
   10 ┤■■■■■■■■■■■■■■■■■■■ 0.08528633700000003   
      └                                        ┘ 
Total (sum): 0.854838688

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks