Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:             153s / 0.28%           1.56GiB / 3.30%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    428ms   100%   214ms   52.3MiB  99.3%  26.1MiB
   applying f to su...  2.00k   40.1ms  9.38%  20.1μs   4.12MiB  7.83%  2.11KiB
   submatrix constr...  2.00k   9.46ms  2.21%  4.73μs   2.18MiB  4.14%  1.12KiB
   writing result c...  2.00k    308μs  0.07%   154ns     0.00B  0.00%    0.00B
 allocate result            2   48.5μs  0.01%  24.3μs    358KiB  0.67%   179KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.111819231
2: 0.680148159
3: 0.704641468
4: 0.699274093
5: 0.646581955
6: 0.685114803
7: 0.646018527
8: 0.647100299
9: 0.688561445
10: 0.710708581

Total (sum): 6.219968561
Std (σ): 0.18091152744362177

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.074930585   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.07705315   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.074908874   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.076187964  
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.074409592   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.075090211   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.074077597   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.076939439  
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.075990254  
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.075150391   
      └                                        ┘ 
Total (sum): 0.7547380570000001
Std (σ): 0.0010201894356705477
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165213622   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165846474   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.164834947   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165925404   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.166643461   
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.828463908
Std (σ): 0.0006976747056496253

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1111     
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1137    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1074      
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1113     
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1080      
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1106     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1183   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1126     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1094      
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▎ 6                                       
   ( 70.0,  75.0] ┤▋ 39                                      
   ( 75.0,  80.0] ┤██▍ 135                                   
   ( 80.0,  85.0] ┤███████▏ 399                              
   ( 85.0,  90.0] ┤███████████████▋ 890                      
   ( 90.0,  95.0] ┤███████████████████████████▍ 1549         
   ( 95.0, 100.0] ┤██████████████████████████████████  1927  
   (100.0, 105.0] ┤█████████████████████████████████▎ 1879   
   (105.0, 110.0] ┤██████████████████████████▊ 1521          
   (110.0, 115.0] ┤████████████████▍ 927                     
   (115.0, 120.0] ┤████████▍ 477                             
   (120.0, 125.0] ┤██▉ 170                                   
   (125.0, 130.0] ┤█▏ 59                                     
   (130.0, 135.0] ┤▍ 17                                      
   (135.0, 140.0] ┤▏ 3                                       
   (140.0, 145.0] ┤▏ 1                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks