Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:        18446452243s / 0.00%        1.48GiB / 4.21%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    455ms   100%   228ms   63.4MiB  99.5%  31.7MiB
   applying f to su...  2.00k   13.8ms  3.04%  6.92μs   13.2MiB  20.7%  6.76KiB
   submatrix constr...  2.00k   11.0ms  2.41%  5.49μs   2.12MiB  3.33%  1.09KiB
   writing result c...  2.00k    353μs  0.08%   176ns     0.00B  0.00%    0.00B
 allocate result            2   99.1μs  0.02%  49.6μs    354KiB  0.54%   177KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a thread_timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(thread_timers=true));
Individual thread timings

1: 0.099453553
2: 0.11058776399999996
3: 0.10908805099999998
4: 0.10379306600000002
5: 0.107192292
6: 0.10632190500000001
7: 0.10532574799999998
8: 0.10197666499999997
9: 0.09789748399999995
10: 0.10081699199999998

Total (sum): 1.0424535199999998

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(thread_timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■ 0.091979138           
    2 ┤■■■■■■■■■■■■■■■■■■■ 0.09292436199999998   
    3 ┤■■■■■■■■■■■■■■■■■■■ 0.09269890800000001   
    4 ┤■■■■■■■■■■■■■■■■■■■ 0.092831704           
    5 ┤■■■■■■■■■■■■■■■■■■■ 0.09307392600000004   
    6 ┤■■■■■■■■■■■■■■■■■■■ 0.09320833599999999   
    7 ┤■■■■■■■■■■■■■■■■■■■ 0.09265931899999998   
    8 ┤■■■■■■■■■■■■■■■■■■■ 0.09292284599999998   
    9 ┤■■■■■■■■■■■■■■■■■■■ 0.0928507             
   10 ┤■■■■■■■■■■■■■■■■■■■ 0.09261122500000002   
      └                                        ┘ 
Total (sum): 0.9277604640000001

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks