Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:             579s / 0.07%           1.59GiB / 3.26%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    409ms   100%   204ms   52.8MiB  99.3%  26.4MiB
   applying f to su...  2.00k   23.5ms  5.75%  11.8μs   4.38MiB  8.23%  2.24KiB
   submatrix constr...  2.00k   9.06ms  2.22%  4.53μs   2.27MiB  4.27%  1.16KiB
   writing result c...  2.00k    319μs  0.08%   159ns     0.00B  0.00%    0.00B
 allocate result            2   44.0μs  0.01%  22.0μs    367KiB  0.67%   183KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.066412077
2: 0.565737297
3: 0.577842946
4: 0.558011709
5: 0.563704132
6: 0.5809005
7: 0.569642059
8: 0.570771003
9: 0.587083188
10: 0.550981735

Total (sum): 5.1910866460000005
Std (σ): 0.15942174280149332

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.090973793   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.089110472    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.088510825    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.08941194    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.087014277    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.087453891    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.08897846     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.090790577   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.089935278   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.086332384    
      └                                        ┘ 
Total (sum): 0.888511897
Std (σ): 0.0015501506542511022
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.149933853   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.145654305    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.141660711    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.147103054    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.145330451    
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.729682374
Std (σ): 0.003002872800307113

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1102     
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1113    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1086     
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1160   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1096     
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1122    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1151   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1093     
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1086     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1137    
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1115   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1113   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1114   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▎ 6                                       
   ( 70.0,  75.0] ┤▌ 25                                      
   ( 75.0,  80.0] ┤██▋ 152                                   
   ( 80.0,  85.0] ┤███████▍ 414                              
   ( 85.0,  90.0] ┤██████████████▋ 836                       
   ( 90.0,  95.0] ┤█████████████████████████▊ 1477           
   ( 95.0, 100.0] ┤██████████████████████████████████  1940  
   (100.0, 105.0] ┤█████████████████████████████████▌ 1913   
   (105.0, 110.0] ┤██████████████████████████▊ 1534          
   (110.0, 115.0] ┤█████████████████▎ 982                    
   (115.0, 120.0] ┤███████▌ 424                              
   (120.0, 125.0] ┤███▋ 205                                  
   (125.0, 130.0] ┤█▎ 66                                     
   (130.0, 135.0] ┤▍ 22                                      
   (135.0, 140.0] ┤▏ 3                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.print_timings

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_timings — Method

Pretty print the given timings.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks