Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Warning

This form of benchmarking doesn't currently work for Threaded()and you'll likely get lot's of errors. See Thread Timers for an alternative.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:           59103s / 0.00%           1.76GiB / 0.39%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2   35.9ms   100%  17.9ms   6.71MiB  94.9%  3.35MiB
   applying f to su...  2.00k   22.0ms  61.1%  11.0μs   4.42MiB  62.5%  2.26KiB
   submatrix constr...  2.00k   11.5ms  32.0%  5.76μs   2.29MiB  32.4%  1.17KiB
   writing result c...  2.00k    287μs  0.80%   143ns     0.00B  0.00%    0.00B
 allocate result            2   57.6μs  0.16%  28.8μs    368KiB  5.09%   184KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Note

You can safely call all *_benchmarks functions in combination with @everywhere, i.e. when using multiple Julia workers (Distributed.jl).

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.119660549
2: 0.693992159
3: 0.709054019
4: 0.685061203
5: 0.682621138
6: 0.692030193
7: 0.67593926
8: 0.711268765
9: 0.684486596
10: 0.681657896

Total (sum): 6.335771778
Std (σ): 0.18094313035101306

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.103112616    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.104248465    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.107840019   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.106975905   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.104104963    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.103262183    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.102058744    
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.104222501    
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.101376863     
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.101614532     
      └                                        ┘ 
Total (sum): 1.038816791
Std (σ): 0.002141017965685343
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.167851267   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.170627791   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.169116311   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.16864772    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.163453793    
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.8396968819999999
Std (σ): 0.0027038269394764887

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1143   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1081     
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1108    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1128    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1126    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1108    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1123    
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1150   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1127    
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1104    
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1120   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1119   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1119   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▏ 4                                       
   ( 70.0,  75.0] ┤▋ 37                                      
   ( 75.0,  80.0] ┤██▌ 144                                   
   ( 80.0,  85.0] ┤███████▎ 413                              
   ( 85.0,  90.0] ┤███████████████▊ 911                      
   ( 90.0,  95.0] ┤██████████████████████████▏ 1490          
   ( 95.0, 100.0] ┤██████████████████████████████████  1948  
   (100.0, 105.0] ┤█████████████████████████████████▊ 1943   
   (105.0, 110.0] ┤█████████████████████████▎ 1441           
   (110.0, 115.0] ┤████████████████▋ 952                     
   (115.0, 120.0] ┤███████▊ 451                              
   (120.0, 125.0] ┤███▎ 180                                  
   (125.0, 130.0] ┤█▎ 65                                     
   (130.0, 135.0] ┤▍ 15                                      
   (135.0, 140.0] ┤▎ 5                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.print_timings

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_timings — Method

Pretty print the given timings.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks