Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:             161s / 0.38%           1.55GiB / 3.47%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    614ms   100%   307ms   54.7MiB  99.4%  27.3MiB
   applying f to su...  2.00k    183ms  29.7%  91.3μs   4.30MiB  7.81%  2.20KiB
   submatrix constr...  2.00k   11.1ms  1.81%  5.55μs   2.25MiB  4.09%  1.15KiB
   writing result c...  2.00k    322μs  0.05%   161ns     0.00B  0.00%    0.00B
 allocate result            2    107μs  0.02%  53.3μs    365KiB  0.65%   183KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.132780802
2: 0.681984724
3: 0.718592755
4: 0.729241331
5: 0.710585904
6: 0.717240306
7: 0.707824101
8: 0.726806836
9: 0.681470168
10: 0.703914454

Total (sum): 6.510441381
Std (σ): 0.1828308084705239

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.096721199    
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.095690638    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.098359624   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.097535763   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.097591751   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.094771221    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.099117874   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.098533699   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.095802916    
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.096126282    
      └                                        ┘ 
Total (sum): 0.9702509669999998
Std (σ): 0.0014249023764057245
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.166140556     
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.176698108   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.172384094    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.177266385   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.171505044    
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.863994187
Std (σ): 0.004509487704318622

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1027      
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1118    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1134   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1113    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1140   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1075     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1069     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1081     
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1124   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1119    
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1100   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 60.0,  70.0] ┤▏ 8                                       
   ( 70.0,  80.0] ┤█▊ 179                                    
   ( 80.0,  90.0] ┤████████████▎ 1239                        
   ( 90.0, 100.0] ┤██████████████████████████████████  3454  
   (100.0, 110.0] ┤█████████████████████████████████▎ 3366   
   (110.0, 120.0] ┤██████████████▌ 1479                      
   (120.0, 130.0] ┤██▌ 253                                   
   (130.0, 140.0] ┤▎ 20                                      
   (140.0, 150.0] ┤▏ 2                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks