Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Warning

This form of benchmarking doesn't currently work for Threaded()and you'll likely get lot's of errors. See Thread Timers for an alternative.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:           25718s / 0.00%           1.64GiB / 0.41%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2   36.1ms   100%  18.0ms   6.47MiB  94.8%  3.23MiB
   applying f to su...  2.00k   22.8ms  63.0%  11.4μs   4.25MiB  62.3%  2.17KiB
   submatrix constr...  2.00k   10.9ms  30.1%  5.46μs   2.22MiB  32.5%  1.13KiB
   writing result c...  2.00k    302μs  0.83%   151ns     0.00B  0.00%    0.00B
 allocate result            2    116μs  0.32%  57.8μs    362KiB  5.19%   181KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Note

You can safely call all *_benchmarks functions in combination with @everywhere, i.e. when using multiple Julia workers (Distributed.jl).

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 0.076366231
2: 0.667266003
3: 0.681608756
4: 0.641311689
5: 0.645847139
6: 0.653145629
7: 0.671319227
8: 0.643230525
9: 0.673710734
10: 0.649961602

Total (sum): 6.003767535000001
Std (σ): 0.18465987560251096

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.115802343   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.11660298    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.115112665   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.114795091   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.114694994   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.11491364    
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.116524794   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.114975835   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.116634221   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.11395366     
      └                                        ┘ 
Total (sum): 1.154010223
Std (σ): 0.0009340095461872456
julia> submatrix_apply(inv, A, Threaded(timers=true, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.168781849   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.164072263    
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.164119523    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165646765    
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.165993899   
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 0.828614299
Std (σ): 0.0019191030641060567

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1085      
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1107     
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1088      
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1056       
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1180   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1111     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1094     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1048       
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1137    
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1074      
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));
              Submatrix Weight per Thread        
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1099   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1096   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1099   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1098   
      └                                        ┘ 
               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 65.0,  70.0] ┤▎ 6                                       
   ( 70.0,  75.0] ┤▌ 32                                      
   ( 75.0,  80.0] ┤██▍ 136                                   
   ( 80.0,  85.0] ┤██████▌ 380                               
   ( 85.0,  90.0] ┤██████████████▊ 871                       
   ( 90.0,  95.0] ┤█████████████████████████▉ 1517           
   ( 95.0, 100.0] ┤██████████████████████████████████  1988  
   (100.0, 105.0] ┤████████████████████████████████▎ 1883    
   (105.0, 110.0] ┤██████████████████████████▋ 1563          
   (110.0, 115.0] ┤███████████████▋ 915                      
   (115.0, 120.0] ┤███████▍ 424                              
   (120.0, 125.0] ┤███▍ 194                                  
   (125.0, 130.0] ┤█▎ 72                                     
   (130.0, 135.0] ┤▍ 15                                      
   (135.0, 140.0] ┤▏ 2                                       
   (140.0, 145.0] ┤▏ 2                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.print_timings

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_timings — Method

Pretty print the given timings.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks