Benchmarking

Instrumented Profiling (TimerOutputs.jl)

By default, the benchmarking facilities are turned off entirely to avoid any performance overhead.

Note

For now, only the serial variant, i.e. with launch configuration Serial() (default), can be benchmarked in this way. For Threaded(), you'll likely get lot's of errors. Check out Thread Timers instead.

Enable the built-in time measurements with SubmatrixMethod.enable_benchmarks().

Run the functionality that you want to benchmark.

Call SubmatrixMethod.print_benchmarks() to see the timing results.

julia> using SubmatrixMethod
julia> SubmatrixMethod.enable_benchmarks()[ Info: Benchmarks enabled!
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A);
julia> submatrix_apply(inv, A);
julia> SubmatrixMethod.print_benchmarks() ──────────────────────────────────────────────────────────────────────────────
                                       Time                   Allocations      
                               ──────────────────────   ───────────────────────
       Tot / % measured:            1162s / 0.04%           1.54GiB / 3.45%    

 Section               ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────
 submatrix_computat...      2    456ms   100%   228ms   54.2MiB  99.4%  27.1MiB
   applying f to su...  2.00k   36.6ms  8.03%  18.3μs   3.91MiB  7.16%  2.00KiB
   submatrix constr...  2.00k   10.5ms  2.30%  5.24μs   2.17MiB  3.98%  1.11KiB
   writing result c...  2.00k    292μs  0.06%   146ns     0.00B  0.00%    0.00B
 allocate result            2   48.3μs  0.01%  24.2μs    359KiB  0.64%   179KiB
 ──────────────────────────────────────────────────────────────────────────────[ Info: Benchmarks disabled!

Benchmarks can be reset by calling SubmatrixMethod.reset_benchmarks() or turned off again via SubmatrixMethod.disable_benchmarks().

Thread Timers

As an attempt to assess the load-balancing when using multiple threads, Threaded() has a timers option (default: false) which toggles time measurements on individual threads.

julia> using SubmatrixMethod
julia> A = sprandsymposdef(1000, 0.1);
julia> submatrix_apply(inv, A, Threaded(timers=true));
Individual thread timings

1: 1.015569016
2: 1.112951009
3: 1.121017357
4: 1.02580273
5: 1.083710942
6: 1.125828245
7: 1.127926302
8: 1.120330082
9: 1.131030731
10: 1.107967839

Total (sum): 10.972134253

If you also load the UnicodePlots.jl package you'll automatically get a nice visual output instead.

julia> using UnicodePlots
julia> submatrix_apply(inv, A, Threaded(timers=true));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■ 0.696171842      
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.792858943   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.750325079    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.74857509      
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 0.749938361    
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.731846811     
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 0.730615775     
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■ 0.700976085      
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■ 0.670437013       
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■ 0.672443536       
      └                                        ┘ 
Total (sum): 7.244188535
julia> submatrix_apply(inv, A, Threaded(timers=true, mode=:batch, nthreads=5));
               Individual thread timings         
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■ 1.364808883   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 1.284859244     
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■ 1.290345548    
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 1.270448008     
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■ 1.281127995     
    6 ┤ 0.0                                      
    7 ┤ 0.0                                      
    8 ┤ 0.0                                      
    9 ┤ 0.0                                      
   10 ┤ 0.0                                      
      └                                        ┘ 
Total (sum): 6.491589678

Note that UnicodePlots.jl needs to be installed separately and must be loaded after using SubmatrixMethod.

You can always (re-)print the timings of the last run via SubmatrixMethod.print_thread_timers(). To access the values themselves, they are stored in SubmatrixMethod.thread_timers[].

Distribution of Submatrices / Workload

To see how the submatrices have been divided among threads, use the show_distribution option of the Threaded() launch configuration. As for the thread timers above, you can readily get bar plot visualizations just by using UnicodePlots.

julia> using SubmatrixMethod, UnicodePlots
julia> A = sprandsymposdef(1000, 0.01);
julia> submatrix_apply(inv, A, Threaded(show_distribution=true));

               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

We can set balance=true to enable balancing w.r.t. the total weight (sizes) of the submatrices each thread has to handle.

julia> submatrix_apply(inv, A, Threaded(balance=true, show_distribution=true));

               # Submatrices per Thread          
      ┌                                        ┐ 
    1 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    2 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    3 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    4 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    5 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    6 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    7 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    8 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
    9 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
   10 ┤■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 100   
      └                                        ┘

Note that the distribution of the sizes of the submatrices can we visualized with SubmatrixMethod.submatrix_sizes_hist.

julia> A = sprandsymposdef(10_000, 0.01);
julia> SubmatrixMethod.submatrix_sizes_hist(A);
                                Submatrix Sizes              
                  ┌                                        ┐ 
   ( 55.0,  60.0] ┤▏ 1                                       
   ( 60.0,  65.0] ┤▏ 1                                       
   ( 65.0,  70.0] ┤▎ 6                                       
   ( 70.0,  75.0] ┤▋ 36                                      
   ( 75.0,  80.0] ┤██▋ 149                                   
   ( 80.0,  85.0] ┤███████▎ 400                              
   ( 85.0,  90.0] ┤████████████████▌ 923                     
   ( 90.0,  95.0] ┤██████████████████████████▊ 1505          
   ( 95.0, 100.0] ┤██████████████████████████████████  1909  
   (100.0, 105.0] ┤████████████████████████████████▊ 1843    
   (105.0, 110.0] ┤███████████████████████████▎ 1527         
   (110.0, 115.0] ┤█████████████████▏ 955                    
   (115.0, 120.0] ┤████████▌ 476                             
   (120.0, 125.0] ┤███▎ 180                                  
   (125.0, 130.0] ┤█▎ 61                                     
   (130.0, 135.0] ┤▌ 24                                      
   (135.0, 140.0] ┤▏ 4                                       
                  └                                        ┘ 
                                   Frequency

References

Index

SubmatrixMethod.disable_benchmarks

SubmatrixMethod.enable_benchmarks

SubmatrixMethod.print_benchmarks

SubmatrixMethod.print_submatrix_distribution

SubmatrixMethod.print_thread_timers

SubmatrixMethod.reset_benchmarks

SubmatrixMethod.reset_thread_timers

Functions

SubmatrixMethod.disable_benchmarks — Method

disable_benchmarks()

Disables benchmarking.

See: enable_benchmarks

SubmatrixMethod.enable_benchmarks — Method

enable_benchmarks()

Enables benchmarking. This affects all SubmatrixMethod.@bench macro applications.

Results can be printed via SubmatrixMethod.print_benchmarks() and reset via SubmatrixMethod.reset_benchmarks().

See: disable_benchmarks

SubmatrixMethod.print_benchmarks — Method

Print benchmark results.

See: enable_benchmarks

SubmatrixMethod.print_submatrix_distribution — Method

Print the distribution of submatrices and their weights.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.print_thread_timers — Method

Print the current state of the thread timers.

If UnicodePlots.jl is loaded (after using SubmatrixMethod), this function will produce a bar plot.

SubmatrixMethod.reset_benchmarks — Method

Reset benchmark results.

SubmatrixMethod.reset_thread_timers — Method

Reset the thread timers to zero.

SubmatrixMethod.@bench — Macro

Usage: @bench "some description" 3+3

See: enable_benchmarks